- Refactored file structure to increase expendability. Added JSDoc style comments to functions
