+++ src/transform/convert/common/generate2xx.js
@@ -1,3 +1,8 @@
+/**
+ * Generates field 245 ($a) based on dc.title values
+ * @param {Object} ValueInterface containing getFields and getFieldValues functions
+ * @returns Empty array or array containing field 245 ($a)
+ */
 export function generate245({getFields, getFieldValues}) {
   const isAddedEntry = generateIsAddedEntry();
   const titles = getFieldValues('dc.title');
@@ -18,6 +23,11 @@ export function generate245({getFields, getFieldValues}) {
   }
 }
 
+/**
+ * Generates field 246 ($a) based on dc.title.alternative values.
+ * @param {Object} ValueInterface containing getFieldValues function
+ * @returns Empty array or array containing field 245 ($a)
+ */
 export function generate246({getFieldValues}) {
   const values = getFieldValues('dc.title.alternative');
   return values.map(value => ({
@@ -26,6 +36,11 @@ export function generate246({getFieldValues}) {
   }));
 }
 
+/**
+ * Generates field 250 ($a) based on dc.description.edition values.
+ * @param {Object} ValueInterface containing getFieldValues function
+ * @returns Empty array or array containing field 250 ($a)
+ */
 export function generate250({getFieldValues}) {
   const values = getFieldValues('dc.description.edition');
   return values.map(value => ({
@@ -34,7 +49,12 @@ export function generate250({getFieldValues}) {
   }));
 }
 
-
+/**
+ * Generates field 264 if any of subfields can be created ($a: optional, $b: optional, $c: optional)
+ * Values used for generation are based on dc.publisher.place, dc.publisher and dc.date.issued.
+ * @param {Object} ValueInterface containing getFieldValues function
+ * @returns Undefined or Object containing field 264 ($a, $b, $c)
+ */
 export function generate264({getFieldValues}) {
   const subfields = generateSubfields();
