+++ src/transform/convert/common/generate7xx.js
@@ -1,5 +1,11 @@
 // Note: Generation of field 700 is defined together with generation of field 100
+// Both fields (100/700) consider author information
 
+/**
+ * Generates field 776 ($z, $9) based on dc.relation.isversionof values.
+ * @param {Object} ValueInterface containing getFieldValues function
+ * @returns Empty array or array containing field 776 ($z, $9)
+ */
 export function generate776({getFieldValues}) {
   const values = getFieldValues('dc.relation.isversionof');
   return values.map(value => ({
