+++ src/transform/filter/filterByIsbnIdentifier.js
@@ -1,5 +1,11 @@
 import {Error as NotSupportedError} from '@natlibfi/melinda-commons';
 
+/**
+ * Filter filtering items without isbn identifiers.
+ * @param {Object} ValueInterface containing getFieldValues function
+ * @param {Object} options Options containing info whether filter should be used
+ * @returns Error if ISBN could not be found from dc.identifier.isbn, otherwise undefined
+ */
 export function filterByIsbnIdentifier({getFieldValues}, options = {}) {
   if (!options.filterByIsbnIdentifier) {
     return;
