+++ src/transform/filter/filterByIssuedYear.js
@@ -1,5 +1,11 @@
 import {Error as NotSupportedError} from '@natlibfi/melinda-commons';
 
+/**
+ * Filter filtering items based on issued year.
+ * @param {Object} ValueInterface containing getFieldValues function
+ * @param {Object} options Options containing info whether filter should be used
+ * @returns Error if record was issued earlier than what defined in configuration, otherwise undefined
+ */
 export function filterByIssuedYear({getFieldValues}, options = {}) {
   if (!options.filterByIssuedYear || options.filterByIssuedYear === 0) {
     return;
